+++ packages/editor/adf-utils/src/validator/specs/block_content.ts
@@ -15,6 +15,7 @@ export default [
   'mediaGroup',
   'decisionList',
   'taskList',
+  'nestableTaskList',
   'table',
   'extension',
   'bodiedExtension',