+++ packages/editor/adf-utils/src/validator/specs/doc.ts
@@ -23,6 +23,7 @@ export default {
           'mediaGroup',
           'decisionList',
           'taskList',
+          'nestableTaskList',
           'table',
           'extension',
           'bodiedExtension',