+++ packages/editor/editor-core/src/plugins/analytics/utils.ts
@@ -1,22 +1,57 @@
-import { editorAnalyticsChannel } from './index';
+import {
+  AnalyticsEventPayloadWithChannel,
+  editorAnalyticsChannel,
+} from './index';
 import { CreateUIAnalyticsEvent } from '@atlaskit/analytics-next';
 import { AnalyticsEventPayload } from './types';
-import { Transaction, EditorState } from 'prosemirror-state';
+import { EditorState, Transaction } from 'prosemirror-state';
 import { Command } from '../../types';
 import { InputRuleWithHandler } from '../../utils/input-rules';
 import { analyticsPluginKey } from './plugin';
+import { AnalyticsStep } from './analytics-step';
+import { Step } from 'prosemirror-transform';
 
 export type DispatchAnalyticsEvent = (payload: AnalyticsEventPayload) => void;
 export type HigherOrderCommand = (command: Command) => Command;
 
+function getAnalyticsState(
+  editorState: EditorState,
+): CreateUIAnalyticsEvent | null | undefined {
+  return analyticsPluginKey.getState(editorState) as
+    | CreateUIAnalyticsEvent
+    | null
+    | undefined;
+}
+
 export function addAnalytics(
+  state: EditorState,
   tr: Transaction,
   payload: AnalyticsEventPayload,
-  channel?: string,
+  channel: string = editorAnalyticsChannel,
 ): Transaction {
-  const analyticsMeta = tr.getMeta(analyticsPluginKey) || [];
-  analyticsMeta.push({ payload, channel });
-  return tr.setMeta(analyticsPluginKey, analyticsMeta);
+  const createAnalyticsEvent = getAnalyticsState(state);
+
+  if (createAnalyticsEvent) {
+    const { storedMarks } = tr;
+    tr.step(
+      new AnalyticsStep(
+        createAnalyticsEvent,
+        [
+          {
+            payload,
+            channel,
+          },
+        ],
+        tr.selection.$from.pos, // We need to create the step based on a position, this prevent split history for relative changes.
+      ),
+    );
+    // When you add a new step all the storedMarks are removed it
+    if (storedMarks) {
+      tr.setStoredMarks(storedMarks);
+    }
+  }
+
+  return tr;
 }
 
 export function withAnalytics(
@@ -33,10 +68,10 @@ export function withAnalytics(
           if (payload instanceof Function) {
             const dynamicPayload = payload(state);
             if (dynamicPayload) {
-              dispatch(addAnalytics(tr, dynamicPayload, channel));
+              dispatch(addAnalytics(state, tr, dynamicPayload, channel));
             }
           } else {
-            dispatch(addAnalytics(tr, payload, channel));
+            dispatch(addAnalytics(state, tr, payload, channel));
           }
         }
       },
@@ -66,7 +101,7 @@ export function ruleWithAnalytics(
 
       if (tr) {
         const payload = getPayload(state, match, start, end);
-        tr = addAnalytics(tr, payload);
+        tr = addAnalytics(state, tr, payload);
       }
       return tr;
     };
@@ -85,3 +120,16 @@ export const fireAnalyticsEvent = (
 }) => {
   return createAnalyticsEvent && createAnalyticsEvent(payload).fire(channel);
 };
+
+export function getAnalyticsEventsFromTransaction(
+  tr: Transaction,
+): AnalyticsEventPayloadWithChannel[] {
+  return (tr.steps as Step[])
+    .filter<AnalyticsStep>(
+      (step: Step): step is AnalyticsStep => step instanceof AnalyticsStep,
+    )
+    .reduce<AnalyticsEventPayloadWithChannel[]>(
+      (acc, step) => [...acc, ...step.analyticsEvents],
+      [],
+    );
+}