+++ packages/editor/editor-core/src/plugins/table/index.tsx
@@ -222,7 +222,7 @@ const tablesPlugin = (options?: TablePluginOptions): EditorPlugin => ({
         icon: () => <IconTable label={formatMessage(messages.table)} />,
         action(insert, state) {
           const tr = insert(createTable(state.schema));
-          return addAnalytics(tr, {
+          return addAnalytics(state, tr, {
             action: ACTION.INSERTED,
             actionSubject: ACTION_SUBJECT.DOCUMENT,
             actionSubjectId: ACTION_SUBJECT_ID.TABLE,