+++ packages/editor/editor-mobile-bridge/src/editor/web-to-native/ios-impl.ts
@@ -4,7 +4,7 @@ import NativeBridge from './bridge';
 import { sendToBridge } from '../../bridge-utils';
 
 export default class IosBridge implements NativeBridge {
-  showMentions(query: String) {
+  showMentions(query: string) {
     if (window.webkit && window.webkit.messageHandlers.mentionBridge) {
       window.webkit.messageHandlers.mentionBridge.postMessage({
         name: 'showMentions',
@@ -138,6 +138,15 @@ export default class IosBridge implements NativeBridge {
         name: 'stateChanged',
         canUndo,
         canRedo,
+      });
+    }
+  }
+
+  trackEvent(event: string) {
+    if (window.webkit && window.webkit.messageHandlers.analyticsBridge) {
+      window.webkit.messageHandlers.analyticsBridge.postMessage({
+        name: 'trackEvent',
+        event,
       });
     }
   }