+++ packages/editor/renderer/src/ui/Renderer/style.tsx
@@ -36,6 +36,7 @@ import {
   dateSharedStyle,
   akEditorFullWidthLayoutWidth,
   mediaSingleClassName,
+  tasksAndDecisionsStyles,
 } from '@atlaskit/editor-common';
 import { RendererCssClassName } from '../../consts';
 import { RendererAppearance } from './types';
@@ -220,6 +221,7 @@ export const Wrapper = styled.div<RendererWrapperProps & HTMLAttributes<{}>>`
   ${codeMarkSharedStyles};
   ${shadowSharedStyle};
   ${dateSharedStyle};
+  ${tasksAndDecisionsStyles};
 
   & .UnknownBlock {
     font-family: ${fontFamily()};